refactor: move org settings to app-router
